font: change ass_face_set_size to use sTypo* for mscale
Freetype 2.10 changes its ascender/descender selection to use
sTypoAscender/sTypoDescender/sTypoLineGap if fsSelection bit 7 is
on for a font. If fsSelection bit 7 is on, and both
ascender/descender == sTypoAscender/sTypoDescender, assume that
Freetype 2.10 or later is in use and set ft_height equal to
sTypo_height for mscale. If Freetype 2.10 or later is not in use
and ascender/descender == sTypoAscender/sTypoDescender, then
hhea Ascender/Descender == sTypoAscender/sTypoDescender, and
ft_height will be equal to sTypo_height anyway.