coretext/fontselect: save index when adding fonts for reuse later
When adding fonts through CoreText, ass_font_provider_add_font adds
fonts with an index of -1, forcing a postscript name lookup when
the font is actually used if it is part of a collection.

This commit adds index to the metadata, so when the font index is
looked up when the font is added, that index can be kept for
future use. This eliminates the need for postscript name lookup
code in ass_font.c.